+++ .idea/modules.xml
@@ -3,6 +3,7 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/LocationTracker.iml" filepath="$PROJECT_DIR$/LocationTracker.iml" />
+      <module fileurl="file://$PROJECT_DIR$/ReminderApp.iml" filepath="$PROJECT_DIR$/ReminderApp.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
