+++ app/src/main/AndroidManifest.xml
@@ -30,7 +30,7 @@
         </activity>
 
         <activity
-            android:name=".Reminder"
+            android:name=".ReminderActivity"
             android:label="@string/title_activity_reminder"
             android:theme="@style/AppTheme.NoActionBar"></activity>
     </application>
