+++ app/src/main/java/com/example/suyog/locationtracker/Reminder.java
@@ -1,147 +1,71 @@
 package com.example.suyog.locationtracker;
 
-import android.app.ProgressDialog;
-import android.content.Context;
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v4.app.FragmentTransaction;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.support.design.widget.NavigationView;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.widget.LinearLayout;
-import android.widget.TextView;
+import java.util.Date;
 
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.ValueEventListener;
+/**
+ * Created by SUYOG on 9/9/2017.
+ */
 
-public class Reminder extends AppCompatActivity
-        implements NavigationView.OnNavigationItemSelectedListener {
+public class Reminder
+{
+    String key;
+    String remianderName;
+    String remainderDate;
+    String remainderTime;
+    String placename;
+    String placeaddress;
+    Double latitude;
+    Double logitude;
 
-    ProgressDialog progressDialog;
-    FirebaseAuth auth;
-
-    TextView name;
-    TextView email;
-    NavigationView navigationView;
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_reminder);
-
-        auth=FirebaseAuth.getInstance();
-        DatabaseReference userReference= FirebaseDatabase.getInstance().getReference("appusers").child(auth.getCurrentUser().getUid());
-        userReference.addValueEventListener(new ValueEventListener() {
-                                                @Override
-                                                public void onDataChange(DataSnapshot dataSnapshot) {
-                                                    Users user = dataSnapshot.getValue(Users.class);
-
-                                                    navigationView = (NavigationView) findViewById(R.id.nav_view);
-                                                    View headerLayout = navigationView.getHeaderView(0);
-
-                                                    name = (TextView) headerLayout.findViewById(R.id.usernametxt);
-                                                    email = (TextView) headerLayout.findViewById(R.id.useremailtxt);
-                                                    name.setText(user.getName().toString());
-                                                    email.setText(auth.getCurrentUser().getEmail());
-                                                }
-
-                                                @Override
-                                                public void onCancelled(DatabaseError databaseError) {
-
-                                                }
-                                            });
-        getSupportFragmentManager().beginTransaction().add(R.id.content,new AddRemainder(),"Add Remainder").commit();
-
-        progressDialog =new ProgressDialog(this);
-
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-
-
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
-        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
-                this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
-        drawer.setDrawerListener(toggle);
-        toggle.syncState();
-
-        NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
-        navigationView.setNavigationItemSelectedListener(this);
+    public Reminder(String key,String remianderName, String remainderDate, String remainderTime, String placename, String placeaddress, Double latitude, Double logitude) {
+        this.key = key;
+        this.remianderName = remianderName;
+        this.remainderDate = remainderDate;
+        this.remainderTime = remainderTime;
+        this.placename = placename;
+        this.placeaddress = placeaddress;
+        this.latitude = latitude;
+        this.logitude = logitude;
     }
 
-    @Override
-    public void onBackPressed() {
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
-        if (drawer.isDrawerOpen(GravityCompat.START)) {
-            drawer.closeDrawer(GravityCompat.START);
-        } else {
-            super.onBackPressed();
-        }
+    public Reminder() {
     }
 
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.remainder, menu);
-        return true;
+    public String getKey()
+    {
+        return key;
     }
 
+    public void setKey(String key)
+    {
+        this.key = key;
+    }
 
-    @SuppressWarnings("StatementWithEmptyBody")
-    @Override
-    public boolean onNavigationItemSelected(MenuItem item) {
-        // Handle navigation view item clicks here.
-        int id = item.getItemId();
+    public String getRemianderName() {
+        return remianderName;
+    }
 
-        if (id == R.id.setremainder) {
-            // Handle the camera action
-            getSupportFragmentManager().beginTransaction().replace(R.id.content,new AddRemainder())
-                    .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE)
-                    .commit();
+    public String getRemainderDate() {
+        return remainderDate;
+    }
 
-        } else if (id == R.id.viewremainder) {
+    public String getRemainderTime() {
+        return remainderTime;
+    }
 
-            getSupportFragmentManager().beginTransaction().replace(R.id.content,new ViewReminder())
-                    .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE)
-                    .commit();
+    public String getPlacename() {
+        return placename;
+    }
 
-        } else if (id == R.id.myprofile) {
+    public String getPlaceaddress() {
+        return placeaddress;
+    }
 
-            getSupportFragmentManager().beginTransaction().replace(R.id.content,new MyProfile())
-                    .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE)
-                    .commit();
+    public Double getLatitude() {
+        return latitude;
+    }
 
-        } else if (id == R.id.deleteremainder) {
-
-            getSupportFragmentManager().beginTransaction().replace(R.id.content,new DeleteRemainder())
-                    .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE)
-                    .commit();
-
-        } else if (id == R.id.loggingout) {
-
-            progressDialog.setMessage("Signing Out");
-            auth.signOut();
-            finish();
-            Intent intent=new Intent(this ,MainActivity.class);
-            startActivity(intent);
-            //setActionBarTitle("Login");
-
-        }
-
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
-        drawer.closeDrawer(GravityCompat.START);
-        return true;
+    public Double getLogitude() {
+        return logitude;
     }
 }
