+++ app/src/main/res/layout/content_reminder.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.suyog.locationtracker.Reminder">
+    tools:context="com.example.suyog.locationtracker.ReminderActivity">
 
 
     <FrameLayout
