+++ app/src/main/res/layout/nav_header_reminder.xml
@@ -12,7 +12,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_top_margin"
     android:theme="@style/ThemeOverlay.AppCompat.Dark"
-    tools:context="com.example.suyog.locationtracker.Reminder">
+    tools:context="com.example.suyog.locationtracker.ReminderActivity">
 
     <ImageView
         android:id="@+id/imageView"
